--- src/redux/books/booksSlice.js
@@ -1,20 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  bookItems: [],
-};
-
-const bookSlice = createSlice({
-  name: 'books',
-  initialState,
-  reducers: {
-    addBook: () => {
-
-    },
-    removeBook: () => {
-
-    },
-  },
-});
-
-export default bookSlice.reducer;
