+++ src/redux/store.js
@@ -1,8 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
-import bookReducer from './books/booksSlice';
+import bookReducer from './books/bookSlice';
+import categoriesReducer from './categories/categoriesSlice';
 
 const store = configureStore({
-  reducer: bookReducer,
+  reducer: {
+    books: bookReducer,
+    categories: categoriesReducer,
+  },
 });
 
 export default store;
